feat(share): collapse to a single latest-only compact codec (#373, #368)
Drop non-latest share-link reading. The share codec is now a single codec
(serialization/compact.ts, renamed from compact-v2.ts) that encodes and
decodes only the current canonical PresetDocument. urlToDocument reads
v === COMPACT_CODEC_VERSION and refuses every other v (old v1.5 links and
versionless pre-#269 links) with UnsupportedVersionError, which surfaces the
existing invalid-link toast and keeps the current session via init().

Reconcile COMPACT_CODEC_VERSION honesty (resolves #368): the old v: 2 label
covered two incompatible filter shapes (scalar #350, canonical pair #361)
across the unreleased #357 epic. With one codec and one shape, bump to v: 3
so version maps 1:1 to shape; any v: 2 payload is now refused uniformly. No
non-latest link was ever released, so nothing in the wild breaks.

Delete the legacy island: v1.5 serialization/compact.ts, encode.ts,
decode.ts, compact.test.ts, and the share-v1_5-* fixtures. Keep the shared
pattern-codec.ts, compress.ts, default-kits.ts, and dense-preset.ts (still
used by document migration/adoption and dense round-trip tests). The
document/ ladder is untouched: .dh files and library kits still migrate
v1/v1.5/v2.

Re-express the codec round-trip tolerance in canonical units: rename the
url-codec test helper to expectCanonicalRoundTrip and compare each field
against its own canonical quantization epsilon (10^-p in seconds/dB/Hz/etc);
no knob-space conversion remains in the codec tests.

Update docs/preset-versioning.md: share links latest-only, single-codec
namespace table, the v1.5 compact decoder leaves the frozen island, and
#368 marked resolved.
