+++ src/features/preset/lib/serialization/__fixtures__/dense-preset.ts
@@ -3,19 +3,16 @@ import type { PresetFileV1 } from "@/features/preset/types/preset";
 import { createEmptyPattern } from "@/features/sequencer/lib/helpers";
 
 /**
- * Deterministic "dense" knob-space (v1.5) preset used by the share-URL
- * fixture corpus: every channel param, every master param, bpm, swing, a
- * 3-step chain with chainEnabled, and per-step accents/ratchets/flams/
- * nudges/velocities are all off-default, on kit-3.
+ * Deterministic "dense" knob-space preset (legacy PresetFileV1, 0-100 knob
+ * space): every channel param, every master param, bpm, swing, a 3-step chain
+ * with chainEnabled, and per-step accents/ratchets/flams/nudges/velocities are
+ * all off-default, on kit-3.
  *
- * FROZEN: the fixture files in this directory were generated by encoding
- * this preset with the v1.5 share codec. Changing this builder desyncs it
- * from those frozen payloads; regenerate the fixtures if it ever changes
- * deliberately.
- *
- * Built as a legacy PresetFileV1 (0-100 knob space) directly, never from the
- * canonical init() document: the v1.5 codec reads and writes this knob shape,
- * and the fixture must reproduce the exact bytes the pre-flip builder emitted.
+ * A synthetic legacy document, not a frozen wire fixture. It is migrated to a
+ * canonical document to seed the dense share round-trip and KIT_ORDER-
+ * independence tests (`url-codec.test.ts`, `url-codec-reorder.test.ts`) and to
+ * exercise the document migration ladder (`library/adoption.test.ts`). Built
+ * directly in the knob shape so it exercises the v1 -> canonical migration.
  */
 function buildDenseSharePreset(): PresetFileV1 {
   const kit = loadKit("kit-3");
