--- src/features/preset/lib/serialization/__fixtures__/share-v1_5-dense.txt
@@ -1,1 +1,0 @@
-eNqNVsuOozAQ_JWVz7Cy2w023GY12p8YzQEI0USrjKIk-5Ci_Pu222AMwRO4NNhVuMpdAt_EYSdqUe6rVnUAudxhn2NbNXnVaZ2bnWqgL1q7Ryky8UfU6nuRiV9EUfR8OIn67SboDTS4p0lgjMwETWAmrqIu7xkDlGEEaEZAxRA3SBhlB5AbJpBGBmnLIC0Z5MqFsAMUFUOxYCgav6RmKMIAKjSDipJBrpx8IVCBA6j02kvDIPdEoLLy6gl7nBY13oSxDDXIUAsMNaMJ6024QiCrGVQVDKrk_Z2er7xtB96pRtQ3NyUqwMrv8U1I8lvcM_Ep6pxef6ZpKaXb8r2_BUGz7UhF4o5UFVFVoGKgKumo3UilRWGkwkA9j8CRY5mzW1Oqp-XCahiYKInp9uXA0812wTB5xZnXtGBpAwfmJp8JhiBYybnVWChOzNATG5g2sgqR1S_0Tk0h6bHHZ3qj1uDc6TO9U2tg4TTWWazEz-2Mp9LdZFUnErwmO2qQT2G7VTYss9htlR11CGU6xuVKCCE4htgxJnK8pjtqFMhkjFd14zKT3VbdU6d8KBNxNnEYMXAw9lokgryqGJapbLcqjpoEMh3omWJY5DERZfuYQxuYNvZaJpK8KhiXgWyfCo7agjId4hXBuAhiIsPVQwB5Z8I3RgWrJhHhmd6oKUMC24164SGA3Ta9UWtQfhFdtfJ_hJlX-j-6P21DC78JRRdNAV1UkC4qli7xzq8Qr_3npf_28_Dv-vvc01z3QYMv8EO9avfUs7X25E4Cmvbu8pfSQp-BI9tyBxbFZxMgKacPfyg5u7MEDXfuPEK_-Y40GlcbJ50qvaygenTnpfv9P093JlY
