--- src/features/preset/lib/serialization/__fixtures__/share-v1_5-init.txt
@@ -1,1 +1,0 @@
-eNqNkUsKwzAMRO8ya7nEtZ22vkrpIv4UTCGEYroJvnvkXQKBygijz5tBoBUlwcOZ8Mg5JDVOb6esyUkF_lXO0TibQrQxgvCD1xdH-LBk4Los8M-10Vm8CEvtYxT4gTDBr6hdyA-NEI4N1nRS_ycJ8dAgfDm9712uEpd04qL3Lka8tRWTTkyOYvImIPkcM1dlLhVtA3lLfq0
