+++ src/features/preset/lib/serialization/compact.ts
@@ -1,367 +1,615 @@
+import { Pattern } from "@/core/audio/engine/pattern-types";
+import { init } from "@/core/dh";
+import { getKitLoader } from "@/core/dhkit";
 import {
-  Pattern,
-  PatternChain,
-  sanitizeChain,
-} from "@/core/audio/engine/pattern-types";
-import type { KitFile } from "@/features/kit/types/kit";
-import type {
-  LegacyKnobInstrumentData,
-  LegacyKnobInstrumentParams,
-  LegacyKnobMasterChainParams,
-} from "@/features/preset/types/legacy-v1";
-import { legacyCycleToChain } from "@/features/sequencer/lib/chain";
-import { VariationCycle } from "@/features/sequencer/types/sequencer";
-import { PRESET_FILE_VERSION } from "../../document/migrate";
-import { PresetFileV1 } from "../../types/preset";
-import { compactCodeToKitId, kitIdToCompactCode } from "./default-kits";
+  CorruptFieldError,
+  InvalidFileError,
+  presetDocumentSchema,
+  UnknownKitError,
+  UnsupportedVersionError,
+  type PresetDocument,
+} from "@/features/preset/document";
 import {
+  CHAIN_STRING_PATTERN,
   CompactAccents,
   CompactVoice,
   decodeAccents,
   decodeVoices,
   encodeAccents,
   encodeVoices,
+  PACKED_TRIGGER_PATTERN,
   parseChainString,
   stringifyChain,
 } from "./pattern-codec";
 
 /**
- * The v1.5 compact share format: knob-space (0-100) values with single-letter
- * keys, bit-packed pattern data, sparse non-default values, and a positional
- * kit code (index into KIT_ORDER).
+ * The share-link compact codec: the single, latest-only encoding of the
+ * domain-unit PresetDocument (docs/preset-persistence.md, "Share encoding").
  *
- * Versioning: the codec was originally unversioned; the `v` field was
- * introduced with the #269 swing retune, mirroring the .dh file version
- * (PRESET_FILE_VERSION = 1.5). Versionless (pre-#269) payloads are no longer
- * decoded at all: urlToDocument refuses them with UnsupportedVersionError
- * (docs/preset-persistence.md, decision 4), so this module only ever sees
- * `v: 1.5` payloads. New links are written by the v2 document codec
- * (compact-v2.ts); this decoder stays alive alongside the v1.x file readers
- * until the v1.x sunset.
+ * There is exactly ONE share codec. Outbound links always encode this shape;
+ * inbound links are read only at this exact version and every other `v`
+ * (older v1.5 knob-space links and pre-#269 versionless links) is refused
+ * with UnsupportedVersionError (#373). No legacy share decode remains.
+ *
+ * Shape:
+ * - `v: 3` (see COMPACT_CODEC_VERSION below), dispatched on by urlToDocument.
+ * - `k` is the STABLE kit id string ("kit-0"), not a positional index into
+ *   KIT_ORDER (decision 12): inserting or reordering registry kits can never
+ *   repoint an existing link.
+ * - Channel and master params carry DOCUMENT values (seconds, dB, semitones,
+ *   0..1 fractions, and the canonical split filter as a `[sideCode, cutoffHz]`
+ *   pair; decision 15), sparse against the init() document's defaults,
+ *   quantized per the table below.
+ * - Nullable volumes (null = silence, the JSON spelling of -Infinity) are
+ *   encoded as an explicit JSON `null`.
+ * - Decode validates shapes and lengths with the typed errors from
+ *   document/errors.ts and finishes with a strict presetDocumentSchema
+ *   parse, so corrupt payloads fail typed, never with a TypeError.
+ *
+ * Pattern data (bit-packed triggers/ratchets/flams/accents, sparse 0-100
+ * velocities, chain strings) is encoded via pattern-codec.ts; it was already
+ * musical-unit data.
  */
+/**
+ * The one share-codec version. Its label is honest: it denotes exactly the
+ * single canonical payload shape above, and nothing else.
+ *
+ * History (#368): the codec's `v: 2` label covered two mutually incompatible
+ * filter shapes across the unreleased #357 epic - a scalar 0-100 position
+ * (#350) and then the canonical `[sideCode, cutoffHz]` pair (#361) - with no
+ * bump. Collapsing to a single codec (#373) is the moment to make the label
+ * truthful: the sole canonical shape takes a fresh version (3) that was never
+ * used for any other shape, so version -> shape is now 1:1. Any `v: 2`
+ * payload (scalar or canonical) is uniformly refused, never decoded, so the
+ * old ambiguity can never resurface. No non-latest link was ever released, so
+ * nothing in the wild breaks.
+ */
+const COMPACT_CODEC_VERSION = 3;
 
-// Legacy-read island: the delta-defaults are FROZEN at the v1.5 init preset's
-// knob values, not derived from the live init() (which is canonical now). Old
-// links omit any field that equalled these defaults, so they must never drift.
-const DEFAULT_SWING = 0;
-const DEFAULT_VARIATION_CYCLE: VariationCycle | undefined = undefined;
-const DEFAULT_PATTERN_CHAIN = sanitizeChain({
-  steps: [{ variation: 0, repeats: 1 }],
-});
-const DEFAULT_CHAIN_ENABLED = false;
-const DEFAULT_BPM = 100;
-
-const DEFAULT_PARAMS: LegacyKnobInstrumentParams = {
-  decay: 100,
-  filter: 50,
-  volume: 92,
-  pan: 50,
-  tune: 50,
-  solo: false,
-  mute: false,
-};
-
-const DEFAULT_MASTER_CHAIN: LegacyKnobMasterChainParams = {
-  filter: 50,
-  saturation: 0,
-  phaser: 0,
-  reverb: 0,
-  compThreshold: 100,
-  compRatio: 57.14285714285714,
-  compAttack: 50,
-  compMix: 70,
-  masterVolume: 92,
-};
-
-const DEFAULT_CHAIN_STRING = stringifyChain(DEFAULT_PATTERN_CHAIN);
-
-function encodeChain(chain: PatternChain): string | undefined {
-  const encoded = stringifyChain(chain);
-  return encoded === DEFAULT_CHAIN_STRING ? undefined : encoded;
-}
-
-function decodeChainString(chainString?: string): PatternChain {
-  if (!chainString) return DEFAULT_PATTERN_CHAIN;
-  return parseChainString(chainString);
-}
-
-// --- COMPACT ENCODING ---
+const CHANNEL_COUNT = 8;
 
 /**
- * Compact instrument params (only non-default values)
+ * Per-field quantization (decimal places written to the payload).
+ *
+ * Requirement: decode(encode(doc)) must yield a document whose canonical
+ * knob values differ from the original's by less than 0.05 knob units (half
+ * the 0.1 knob display step).
+ *
+ * Derivation: quantizing to `p` decimals perturbs the domain value by at
+ * most one grid step q = 10^-p (q/2 from rounding; up to q when a
+ * near-default value is omitted as sparse and decodes to the default). The
+ * knob error is that domain perturbation times the slope of the INVERSE
+ * (domain -> knob) mapping, so each field's precision is set by the inverse
+ * mapping's steepest region:
+ *
+ * - Linear fields have a constant inverse slope of 100/span knob units per
+ *   domain unit, so the bound is uniform: err <= q * 100 / span.
+ * - Exponential fields (decay, compAttack; forward domain =
+ *   min + (knob/100)^2 * span) have inverse knob = 100 * sqrt((d - min) /
+ *   span). The FORWARD curve is steepest at knob 100, which means the
+ *   inverse is steepest at the opposite end: d(knob)/d(domain) =
+ *   50 / sqrt((d - min) * span) diverges as d -> min (knob 0), so the
+ *   binding region is the domain floor, not knob 100. Worst case over a
+ *   grid step at the floor: err <= 100 * sqrt(q / span).
+ *
+ * The split filter is carried as a `[sideCode, cutoffHz]` pair (sideCode
+ * 0 = lowpass, 1 = highpass); its widget position is recovered downstream by
+ * the filter descriptor's filterToPosition (position = 49 * sqrt(cutoffHz /
+ * 15000) per side). That inverse is steepest at the cutoff floor, so like the
+ * exponential
+ * fields the bound is worst at cutoffHz -> 0: err <= 49 * sqrt(q / 15000);
+ * p = 3 keeps it at ~0.013 knob units, well under 0.05.
+ *
+ * | field             | key | mapping (inverse slope, worst)       | p | worst knob err |
+ * |-------------------|-----|--------------------------------------|---|----------------|
+ * | decaySeconds      | d   | exp, span 4.995 s; 100*sqrt(q/span)  | 7 | 0.015          |
+ * | filter (cutoffHz) | f   | 49*sqrt(q/15000) at the floor        | 3 | 0.013          |
+ * | volumeDb          | v   | linear, span 50 dB; 2 knob/dB        | 2 | 0.02           |
+ * | pan               | p   | linear, span 2; 50 knob/unit         | 4 | 0.005          |
+ * | tuneSemitones     | t   | linear, span 14 st; 50/7 knob/st     | 3 | 0.008          |
+ * | saturation        | s   | linear, span 1; 100 knob/unit        | 4 | 0.01           |
+ * | phaser            | ph  | linear, span 1; 100 knob/unit        | 4 | 0.01           |
+ * | reverb            | rv  | linear, span 1; 100 knob/unit        | 4 | 0.01           |
+ * | compThresholdDb   | ct  | linear, span 40 dB; 2.5 knob/dB      | 2 | 0.025          |
+ * | compRatio         | cr  | integer 1..8, carried exactly        | - | 0              |
+ * | compAttackSeconds | ca  | exp, span 0.099 s; 100*sqrt(q/span)  | 8 | 0.032          |
+ * | compMix           | cm  | linear, span 1; 100 knob/unit        | 4 | 0.01           |
+ * | masterVolumeDb    | mv  | linear, span 50 dB; 2 knob/dB        | 2 | 0.02           |
+ * | swing (fraction)  | sw  | linear, span 0.375; 266.7 knob/unit  | 4 | 0.027          |
+ * | bpm               | bpm | raw domain value in both spaces      | - | 0              |
+ * | velocities        |     | sparse ints 0-100 (not a knob)       | - | 0.005 velocity |
  */
-type CompactParams = {
-  d?: number; // decay
-  f?: number; // filter
-  v?: number; // volume
-  p?: number; // pan
-  t?: number; // tune
-  s?: number; // solo (1/0)
-  m?: number; // mute (1/0)
-};
+const PRECISION = {
+  decaySeconds: 7,
+  filter: 3,
+  volumeDb: 2,
+  pan: 4,
+  tuneSemitones: 3,
+  saturation: 4,
+  phaser: 4,
+  reverb: 4,
+  compThresholdDb: 2,
+  compAttackSeconds: 8,
+  compMix: 4,
+  masterVolumeDb: 2,
+  swing: 4,
+} as const;
 
 /**
- * Compact preset format
+ * The sparse baseline: the init preset, which is already a canonical document.
  */
+const INIT_DOCUMENT = init();
+
+type Channel = PresetDocument["channels"][number];
+type Master = PresetDocument["master"];
+
+/** Canonical filter as a compact pair: [sideCode (0=lowpass, 1=highpass), cutoffHz]. */
+type CompactFilter = [number, number];
+
+/** Sparse domain-unit channel params; `v: null` spells silence. */
+type CompactChannel = {
+  d?: number; // decaySeconds
+  f?: CompactFilter; // canonical filter [sideCode, cutoffHz]
+  v?: number | null; // volumeDb (null = silence)
+  p?: number; // pan -1..1
+  t?: number; // tuneSemitones -7..7
+  s?: number; // solo (1)
+  m?: number; // mute (1)
+};
+
+/** Sparse domain-unit master params; `mv: null` spells silence. */
+type CompactMaster = {
+  f?: CompactFilter; // canonical filter [sideCode, cutoffHz]
+  s?: number; // saturation 0..1
+  ph?: number; // phaser 0..1
+  rv?: number; // reverb 0..1
+  ct?: number; // compThresholdDb
+  cr?: number; // compRatio 1..8 integer
+  ca?: number; // compAttackSeconds
+  cm?: number; // compMix 0..1
+  mv?: number | null; // masterVolumeDb (null = silence)
+};
+
 type CompactPreset = {
-  id: string; // preset UUID (new UUID generated when sharing)
-  v?: number; // codec version, mirrors PRESET_FILE_VERSION (absent = refused pre-#269 legacy URL)
-  k: string; // kit code (positional index into KIT_ORDER)
-  n?: string; // preset name
-  ip: CompactParams[]; // instrument params (8 items, only non-defaults)
+  v: 3;
+  id: string; // preset UUID (minted fresh when sharing)
+  n: string; // preset name
+  k: string; // STABLE kit id, e.g. "kit-0" (decision 12)
+  ip: CompactChannel[]; // channel params (8 items, only non-defaults)
   pt: CompactVoice[]; // pattern voices (8 items)
-  ac?: CompactAccents; // accent patterns (hex-encoded, omit if no accents) [A, B, C, D]
-  vc?: string; // legacy variation cycle (omit if undefined)
-  ch?: string; // chain string (variation + repeat pairs)
-  ce?: number; // chain enabled flag
-  bpm?: number; // bpm (omit if 120)
-  sw?: number; // swing (omit if 0)
-  mc?: CompactMasterChain;
+  ac?: CompactAccents; // accent bitmaps [A, B, C, D]
+  ch?: string; // chain string (omit if default "A1")
+  ce?: number; // chain enabled (omit if default)
+  bpm?: number; // omit if init default
+  sw?: number; // swing fraction 0..0.375 (omit if init default)
+  mc?: CompactMaster;
 };
 
-type CompactMasterChain = Partial<{
-  // New format
-  f: number; // filter (unified split filter)
-  s: number; // saturation
-  ph: number; // phaser
-  rv: number; // reverb
-  ct: number; // compThreshold
-  cr: number; // compRatio
-  ca: number; // compAttack
-  cm: number; // compMix
-  mv: number; // masterVolume
+// --- ENCODE ---
 
-  // Legacy format (for backward compatibility during decode)
-  lp?: number; // lowPass (legacy)
-  hp?: number; // highPass (legacy)
-}>;
+function quantize(value: number, decimals: number): number {
+  return Number(value.toFixed(decimals));
+}
 
-// --- ENCODE FUNCTIONS ---
+/** Compact side code: 0 = lowpass, 1 = highpass. */
+function sideCode(side: Channel["filter"]["side"]): number {
+  return side === "highpass" ? 1 : 0;
+}
 
-function encodeParams(params: LegacyKnobInstrumentParams): CompactParams {
-  const compact: CompactParams = {};
+/**
+ * Sparse canonical filter: the `[sideCode, cutoffHz]` pair, or undefined when
+ * it lands on the (quantized) default. cutoffHz is compared quantized so an
+ * omission and a written value round-trip to the same value.
+ */
+function sparseFilter(
+  filter: Channel["filter"],
+  defaultFilter: Channel["filter"],
+): CompactFilter | undefined {
+  const side = sideCode(filter.side);
+  const cutoff = quantize(filter.cutoffHz, PRECISION.filter);
+  if (
+    side === sideCode(defaultFilter.side) &&
+    cutoff === quantize(defaultFilter.cutoffHz, PRECISION.filter)
+  ) {
+    return undefined;
+  }
+  return [side, cutoff];
+}
 
-  if (params.decay !== DEFAULT_PARAMS.decay) compact.d = params.decay;
-  if (params.filter !== DEFAULT_PARAMS.filter) compact.f = params.filter;
-  if (params.volume !== DEFAULT_PARAMS.volume) compact.v = params.volume;
-  if (params.pan !== DEFAULT_PARAMS.pan) compact.p = params.pan;
-  if (params.tune !== DEFAULT_PARAMS.tune) compact.t = params.tune;
-  if (params.solo !== DEFAULT_PARAMS.solo) compact.s = params.solo ? 1 : 0;
-  if (params.mute !== DEFAULT_PARAMS.mute) compact.m = params.mute ? 1 : 0;
+/** Quantized value, or undefined when it lands on the (quantized) default. */
+function sparse(
+  value: number,
+  defaultValue: number,
+  decimals: number,
+): number | undefined {
+  const quantized = quantize(value, decimals);
+  return quantized === quantize(defaultValue, decimals) ? undefined : quantized;
+}
+
+/** sparse() over nullable volumes: null (silence) is spelled explicitly. */
+function sparseNullable(
+  value: number | null,
+  defaultValue: number | null,
+  decimals: number,
+): number | null | undefined {
+  if (value === null) return defaultValue === null ? undefined : null;
+  if (defaultValue === null) return quantize(value, decimals);
+  return sparse(value, defaultValue, decimals);
+}
+
+function encodeChannel(channel: Channel, defaults: Channel): CompactChannel {
+  const compact: CompactChannel = {};
+
+  const d = sparse(
+    channel.decaySeconds,
+    defaults.decaySeconds,
+    PRECISION.decaySeconds,
+  );
+  const f = sparseFilter(channel.filter, defaults.filter);
+  const v = sparseNullable(
+    channel.volumeDb,
+    defaults.volumeDb,
+    PRECISION.volumeDb,
+  );
+  const p = sparse(channel.pan, defaults.pan, PRECISION.pan);
+  const t = sparse(
+    channel.tuneSemitones,
+    defaults.tuneSemitones,
+    PRECISION.tuneSemitones,
+  );
+
+  if (d !== undefined) compact.d = d;
+  if (f !== undefined) compact.f = f;
+  if (v !== undefined) compact.v = v;
+  if (p !== undefined) compact.p = p;
+  if (t !== undefined) compact.t = t;
+  if (channel.solo !== defaults.solo) compact.s = channel.solo ? 1 : 0;
+  if (channel.mute !== defaults.mute) compact.m = channel.mute ? 1 : 0;
 
   return compact;
 }
 
-function encodeMasterChain(
-  chain: LegacyKnobMasterChainParams,
-): CompactMasterChain | undefined {
-  const mc: CompactMasterChain = {};
-  if (chain.filter !== DEFAULT_MASTER_CHAIN.filter) mc.f = chain.filter;
-  if (chain.saturation !== DEFAULT_MASTER_CHAIN.saturation)
-    mc.s = chain.saturation;
-  if (chain.phaser !== DEFAULT_MASTER_CHAIN.phaser) mc.ph = chain.phaser;
-  if (chain.reverb !== DEFAULT_MASTER_CHAIN.reverb) mc.rv = chain.reverb;
-  if (chain.compThreshold !== DEFAULT_MASTER_CHAIN.compThreshold)
-    mc.ct = chain.compThreshold;
-  if (chain.compRatio !== DEFAULT_MASTER_CHAIN.compRatio)
-    mc.cr = chain.compRatio;
-  if (chain.compAttack !== DEFAULT_MASTER_CHAIN.compAttack)
-    mc.ca = chain.compAttack;
-  if (chain.compMix !== DEFAULT_MASTER_CHAIN.compMix) mc.cm = chain.compMix;
-  if (chain.masterVolume !== DEFAULT_MASTER_CHAIN.masterVolume)
-    mc.mv = chain.masterVolume;
+function encodeMaster(
+  master: Master,
+  defaults: Master,
+): CompactMaster | undefined {
+  const compact: CompactMaster = {};
 
-  return Object.keys(mc).length > 0 ? mc : undefined;
+  const f = sparseFilter(master.filter, defaults.filter);
+  const s = sparse(
+    master.saturation,
+    defaults.saturation,
+    PRECISION.saturation,
+  );
+  const ph = sparse(master.phaser, defaults.phaser, PRECISION.phaser);
+  const rv = sparse(master.reverb, defaults.reverb, PRECISION.reverb);
+  const ct = sparse(
+    master.compThresholdDb,
+    defaults.compThresholdDb,
+    PRECISION.compThresholdDb,
+  );
+  const ca = sparse(
+    master.compAttackSeconds,
+    defaults.compAttackSeconds,
+    PRECISION.compAttackSeconds,
+  );
+  const cm = sparse(master.compMix, defaults.compMix, PRECISION.compMix);
+  const mv = sparseNullable(
+    master.masterVolumeDb,
+    defaults.masterVolumeDb,
+    PRECISION.masterVolumeDb,
+  );
+
+  if (f !== undefined) compact.f = f;
+  if (s !== undefined) compact.s = s;
+  if (ph !== undefined) compact.ph = ph;
+  if (rv !== undefined) compact.rv = rv;
+  if (ct !== undefined) compact.ct = ct;
+  if (master.compRatio !== defaults.compRatio) compact.cr = master.compRatio;
+  if (ca !== undefined) compact.ca = ca;
+  if (cm !== undefined) compact.cm = cm;
+  if (mv !== undefined) compact.mv = mv;
+
+  return Object.keys(compact).length > 0 ? compact : undefined;
 }
 
-function encodeCompactPreset(preset: PresetFileV1): CompactPreset {
-  const kitId = kitIdToCompactCode(preset.kit.meta.id);
-  if (!kitId) {
-    throw new Error(
-      `Unknown kit ID: ${preset.kit.meta.id}. Only default kits can be shared.`,
-    );
+const DEFAULT_CHAIN_STRING = stringifyChain(INIT_DOCUMENT.playback.chain);
+
+/**
+ * Encode a preset document as the v2 compact share payload.
+ *
+ * @throws {UnknownKitError} If the document's kit id does not resolve in the
+ * registry (nothing in the app can currently produce one)
+ */
+function encodeCompactDocument(document: PresetDocument): CompactPreset {
+  if (getKitLoader(document.kit.id) === undefined) {
+    throw new UnknownKitError(document.kit.id);
   }
 
   const compact: CompactPreset = {
-    id: preset.meta.id, // Include the UUID (generated fresh when sharing)
-    v: PRESET_FILE_VERSION,
-    k: kitId,
-    ip: preset.kit.instruments.map((inst: LegacyKnobInstrumentData) =>
-      encodeParams(inst.params),
+    v: COMPACT_CODEC_VERSION,
+    id: document.meta.id,
+    n: document.meta.name,
+    k: document.kit.id,
+    ip: document.channels.map((channel, index) =>
+      encodeChannel(channel, INIT_DOCUMENT.channels[index]),
     ),
-    pt: encodeVoices(preset.sequencer.pattern),
+    pt: encodeVoices(document.pattern),
   };
 
-  // Encode accent patterns (only if any accents exist)
-  const accents = encodeAccents(preset.sequencer.pattern.variationMetadata);
+  const accents = encodeAccents(document.pattern.variationMetadata);
   if (accents) {
     compact.ac = accents;
   }
 
-  // Always include preset name
-  compact.n = preset.meta.name;
-
-  if (
-    preset.sequencer.variationCycle &&
-    preset.sequencer.variationCycle !== DEFAULT_VARIATION_CYCLE
-  ) {
-    compact.vc = preset.sequencer.variationCycle;
+  const chainString = stringifyChain(document.playback.chain);
+  if (chainString !== DEFAULT_CHAIN_STRING) {
+    compact.ch = chainString;
   }
 
-  const encodedChain = encodeChain(
-    preset.sequencer.chain ?? DEFAULT_PATTERN_CHAIN,
+  if (document.playback.chainEnabled !== INIT_DOCUMENT.playback.chainEnabled) {
+    compact.ce = document.playback.chainEnabled ? 1 : 0;
+  }
+
+  if (document.transport.bpm !== INIT_DOCUMENT.transport.bpm) {
+    compact.bpm = document.transport.bpm;
+  }
+
+  const sw = sparse(
+    document.transport.swing,
+    INIT_DOCUMENT.transport.swing,
+    PRECISION.swing,
   );
-
-  if (encodedChain) {
-    compact.ch = encodedChain;
+  if (sw !== undefined) {
+    compact.sw = sw;
   }
 
-  if (preset.sequencer.chainEnabled !== DEFAULT_CHAIN_ENABLED) {
-    compact.ce = preset.sequencer.chainEnabled ? 1 : 0;
-  }
-
-  if (preset.transport.bpm !== DEFAULT_BPM) {
-    compact.bpm = preset.transport.bpm;
-  }
-
-  if (preset.transport.swing !== DEFAULT_SWING) {
-    compact.sw = preset.transport.swing;
-  }
-
-  const masterChain = encodeMasterChain(preset.masterChain);
-  if (masterChain) {
-    compact.mc = masterChain;
+  const master = encodeMaster(document.master, INIT_DOCUMENT.master);
+  if (master) {
+    compact.mc = master;
   }
 
   return compact;
 }
 
-// --- DECODE FUNCTIONS ---
+// --- DECODE ---
 
-function decodeParams(compact: CompactParams): LegacyKnobInstrumentParams {
+function corrupt(path: string, message: string): never {
+  throw new CorruptFieldError(path, message);
+}
+
+function isRecord(value: unknown): value is Record<string, unknown> {
+  return typeof value === "object" && value !== null && !Array.isArray(value);
+}
+
+function validatePackedFlags(value: unknown, path: string): void {
+  if (typeof value !== "string" || !PACKED_TRIGGER_PATTERN.test(value)) {
+    corrupt(path, "expected a 4-character hex step bitmap");
+  }
+}
+
+function validateStepSequence(value: unknown, path: string): void {
+  if (!isRecord(value)) corrupt(path, "expected a step sequence object");
+
+  validatePackedFlags(value.t, `${path}.t`);
+
+  if (value.v !== undefined) {
+    if (!isRecord(value.v)) corrupt(`${path}.v`, "expected a velocity map");
+    for (const [key, velocity] of Object.entries(value.v)) {
+      const step = Number(key);
+      if (!Number.isInteger(step) || step < 0 || step >= 16) {
+        corrupt(`${path}.v.${key}`, "velocity step index out of range");
+      }
+      if (typeof velocity !== "number" || !Number.isFinite(velocity)) {
+        corrupt(`${path}.v.${key}`, "expected a numeric velocity");
+      }
+    }
+  }
+
+  if (value.n !== undefined && typeof value.n !== "number") {
+    corrupt(`${path}.n`, "expected a numeric timing nudge");
+  }
+  if (value.r !== undefined) validatePackedFlags(value.r, `${path}.r`);
+  if (value.f !== undefined) validatePackedFlags(value.f, `${path}.f`);
+}
+
+function validateVoices(value: unknown): asserts value is CompactVoice[] {
+  if (!Array.isArray(value) || value.length !== CHANNEL_COUNT) {
+    corrupt("pt", `expected ${CHANNEL_COUNT} pattern voices`);
+  }
+  value.forEach((voice, index) => {
+    if (!isRecord(voice)) corrupt(`pt.${index}`, "expected a voice object");
+    if (typeof voice.i !== "number") {
+      corrupt(`pt.${index}.i`, "expected a numeric instrument index");
+    }
+    validateStepSequence(voice.a, `pt.${index}.a`);
+    validateStepSequence(voice.b, `pt.${index}.b`);
+    if (voice.c !== undefined) validateStepSequence(voice.c, `pt.${index}.c`);
+    if (voice.d !== undefined) validateStepSequence(voice.d, `pt.${index}.d`);
+  });
+}
+
+function validateChannels(value: unknown): asserts value is CompactChannel[] {
+  if (!Array.isArray(value) || value.length !== CHANNEL_COUNT) {
+    corrupt("ip", `expected ${CHANNEL_COUNT} channel param objects`);
+  }
+  value.forEach((channel, index) => {
+    if (!isRecord(channel)) {
+      corrupt(`ip.${index}`, "expected a channel params object");
+    }
+  });
+}
+
+function validateAccents(value: unknown): CompactAccents | undefined {
+  if (value === undefined) return undefined;
+  if (!Array.isArray(value) || value.length > 4) {
+    corrupt("ac", "expected up to 4 accent bitmaps");
+  }
+  const accents = value.map((entry, index) => {
+    if (entry === undefined || entry === null) return undefined;
+    validatePackedFlags(entry, `ac.${index}`);
+    return entry as string;
+  });
+  return accents as CompactAccents;
+}
+
+/**
+ * Decode the compact `[sideCode, cutoffHz]` pair to a canonical filter, or
+ * fall back to the default when absent. A malformed pair fails typed (never a
+ * TypeError); the final schema parse still range-checks cutoffHz.
+ */
+function decodeFilter(
+  f: CompactFilter | undefined,
+  defaultFilter: Channel["filter"],
+  path: string,
+): Channel["filter"] {
+  if (f === undefined) return defaultFilter;
+  if (
+    !Array.isArray(f) ||
+    f.length !== 2 ||
+    typeof f[0] !== "number" ||
+    typeof f[1] !== "number"
+  ) {
+    corrupt(path, "expected a [sideCode, cutoffHz] filter pair");
+  }
+  if (f[0] !== 0 && f[0] !== 1) {
+    corrupt(path, "filter side code must be 0 (lowpass) or 1 (highpass)");
+  }
+  return { side: f[0] === 1 ? "highpass" : "lowpass", cutoffHz: f[1] };
+}
+
+function decodeChannel(
+  compact: CompactChannel,
+  defaults: Channel,
+  index: number,
+): Channel {
   return {
-    decay: compact.d ?? DEFAULT_PARAMS.decay,
-    filter: compact.f ?? DEFAULT_PARAMS.filter,
-    volume: compact.v ?? DEFAULT_PARAMS.volume,
-    pan: compact.p ?? DEFAULT_PARAMS.pan,
-    tune: compact.t ?? DEFAULT_PARAMS.tune,
+    decaySeconds: compact.d ?? defaults.decaySeconds,
+    filter: decodeFilter(compact.f, defaults.filter, `ip.${index}.f`),
+    volumeDb: compact.v !== undefined ? compact.v : defaults.volumeDb,
+    pan: compact.p ?? defaults.pan,
+    tuneSemitones: compact.t ?? defaults.tuneSemitones,
     solo: compact.s === 1,
     mute: compact.m === 1,
   };
 }
 
-function decodeMasterChain(
-  compact?: CompactMasterChain,
-): LegacyKnobMasterChainParams {
-  // Handle legacy format (lp/hp) or new format (f/s/ca)
-  if (compact?.f !== undefined) {
-    // New format
-    return {
-      filter: compact.f,
-      saturation: compact.s ?? DEFAULT_MASTER_CHAIN.saturation,
-      phaser: compact.ph ?? DEFAULT_MASTER_CHAIN.phaser,
-      reverb: compact.rv ?? DEFAULT_MASTER_CHAIN.reverb,
-      compThreshold: compact.ct ?? DEFAULT_MASTER_CHAIN.compThreshold,
-      compRatio: compact.cr ?? DEFAULT_MASTER_CHAIN.compRatio,
-      compAttack: compact.ca ?? DEFAULT_MASTER_CHAIN.compAttack,
-      compMix: compact.cm ?? DEFAULT_MASTER_CHAIN.compMix,
-      masterVolume: compact.mv ?? DEFAULT_MASTER_CHAIN.masterVolume,
-    };
-  }
-
-  // Legacy format: convert lp/hp to unified filter
-  // This will be further processed by migrateMasterChainParams
+function decodeMaster(
+  compact: CompactMaster | undefined,
+  defaults: Master,
+): Master {
   return {
-    filter: DEFAULT_MASTER_CHAIN.filter,
-    saturation: DEFAULT_MASTER_CHAIN.saturation,
-    phaser: compact?.ph ?? DEFAULT_MASTER_CHAIN.phaser,
-    reverb: compact?.rv ?? DEFAULT_MASTER_CHAIN.reverb,
-    compThreshold: compact?.ct ?? DEFAULT_MASTER_CHAIN.compThreshold,
-    compRatio: compact?.cr ?? DEFAULT_MASTER_CHAIN.compRatio,
-    compAttack: compact?.ca ?? DEFAULT_MASTER_CHAIN.compAttack,
-    compMix: compact?.cm ?? DEFAULT_MASTER_CHAIN.compMix,
-    masterVolume: compact?.mv ?? DEFAULT_MASTER_CHAIN.masterVolume,
-    // Include legacy values for migration
-    lowPass: compact?.lp,
-    highPass: compact?.hp,
+    filter: decodeFilter(compact?.f, defaults.filter, "mc.f"),
+    saturation: compact?.s ?? defaults.saturation,
+    phaser: compact?.ph ?? defaults.phaser,
+    reverb: compact?.rv ?? defaults.reverb,
+    compThresholdDb: compact?.ct ?? defaults.compThresholdDb,
+    compRatio: compact?.cr ?? defaults.compRatio,
+    compAttackSeconds: compact?.ca ?? defaults.compAttackSeconds,
+    compMix: compact?.cm ?? defaults.compMix,
+    masterVolumeDb:
+      compact?.mv !== undefined ? compact.mv : defaults.masterVolumeDb,
   };
 }
 
-function decodeCompactPreset(
-  compact: CompactPreset,
-  kitLoader: (kitId: string) => KitFile,
-): PresetFileV1 {
-  const kitId = compactCodeToKitId(compact.k);
-  if (!kitId) {
-    throw new Error(`Unknown kit code: ${compact.k}`);
+/**
+ * Decode a parsed v2 compact payload into a validated PresetDocument.
+ *
+ * @throws {InvalidFileError} If the payload is not an object
+ * @throws {UnsupportedVersionError} If the payload's `v` is not the single
+ * current COMPACT_CODEC_VERSION
+ * @throws {UnknownKitError} If the kit id does not resolve in the registry
+ * @throws {CorruptFieldError} If any field fails shape, length, grammar, or
+ * (via the final presetDocumentSchema parse) range validation
+ */
+function decodeCompactDocument(data: unknown): PresetDocument {
+  if (!isRecord(data)) {
+    throw new InvalidFileError("Shared preset payload must be an object");
+  }
+  if (data.v !== COMPACT_CODEC_VERSION) {
+    throw new UnsupportedVersionError(data.v);
   }
 
-  const defaultKit = kitLoader(kitId);
+  if (typeof data.id !== "string") corrupt("id", "expected a preset id string");
+  if (data.k === undefined) throw new UnknownKitError(undefined);
+  if (typeof data.k !== "string") corrupt("k", "expected a kit id string");
+  if (getKitLoader(data.k) === undefined) {
+    throw new UnknownKitError(data.k);
+  }
+  if (data.n !== undefined && typeof data.n !== "string") {
+    corrupt("n", "expected a preset name string");
+  }
+
+  validateChannels(data.ip);
+  validateVoices(data.pt);
+  const accents = validateAccents(data.ac);
+
+  if (data.ch !== undefined) {
+    if (typeof data.ch !== "string" || !CHAIN_STRING_PATTERN.test(data.ch)) {
+      corrupt("ch", "expected a chain string of 1-8 [A-D][1-8] pairs");
+    }
+  }
+  if (data.ce !== undefined && data.ce !== 0 && data.ce !== 1) {
+    corrupt("ce", "expected chain-enabled flag 0 or 1");
+  }
+  if (data.bpm !== undefined && typeof data.bpm !== "number") {
+    corrupt("bpm", "expected a numeric bpm");
+  }
+  if (data.sw !== undefined && typeof data.sw !== "number") {
+    corrupt("sw", "expected a numeric swing fraction");
+  }
+  if (data.mc !== undefined && !isRecord(data.mc)) {
+    corrupt("mc", "expected a master params object");
+  }
 
   const pattern: Pattern = {
-    voices: decodeVoices(compact.pt),
-    // Decode accent patterns (default to no accents if not present)
-    variationMetadata: decodeAccents(compact.ac),
+    voices: decodeVoices(data.pt),
+    variationMetadata: decodeAccents(accents),
   };
 
-  // The registry kit is canonical; the v1.5 decoder keeps its sample/meta
-  // identity and overrides with the decoded knob params (legacy shape).
-  const instruments: LegacyKnobInstrumentData[] = defaultKit.instruments.map(
-    (inst, idx: number) => ({
-      meta: inst.meta,
-      role: inst.role,
-      sample: inst.sample,
-      params: decodeParams(compact.ip[idx]),
-    }),
-  );
+  const now = new Date().toISOString();
 
-  const legacyChain = legacyCycleToChain(
-    (compact.vc ?? DEFAULT_VARIATION_CYCLE) as VariationCycle | undefined,
-    0,
-  );
-
-  const chain = compact.ch
-    ? decodeChainString(compact.ch)
-    : (legacyChain.chain ?? DEFAULT_PATTERN_CHAIN);
-
-  const chainEnabled =
-    compact.ce !== undefined
-      ? compact.ce === 1
-      : (legacyChain.chainEnabled ?? DEFAULT_CHAIN_ENABLED);
-
-  return {
-    kind: "drumhaus.preset",
-    // Decoded presets are normalized to the current knob-space file
-    // version: only v: 1.5 payloads reach this decoder (versionless legacy
-    // URLs are refused upstream, decision 4), so no swing migration applies.
-    version: PRESET_FILE_VERSION,
+  const candidate: PresetDocument = {
+    kind: INIT_DOCUMENT.kind,
+    version: INIT_DOCUMENT.version,
     meta: {
-      id: compact.id, // Use the UUID from the encoded preset
-      name: compact.n || "Shared Preset",
-      createdAt: new Date().toISOString(),
-      updatedAt: new Date().toISOString(),
+      id: data.id,
+      name: data.n || "Shared Preset",
+      createdAt: now,
+      updatedAt: now,
     },
-    kit: {
-      kind: "drumhaus.kit",
-      version: 1,
-      meta: defaultKit.meta,
-      instruments,
+    kit: { id: data.k },
+    channels: data.ip.map((channel, index) =>
+      decodeChannel(channel, INIT_DOCUMENT.channels[index], index),
+    ) as PresetDocument["channels"],
+    pattern,
+    playback: {
+      chain:
+        data.ch !== undefined
+          ? parseChainString(data.ch)
+          : INIT_DOCUMENT.playback.chain,
+      chainEnabled:
+        data.ce !== undefined
+          ? data.ce === 1
+          : INIT_DOCUMENT.playback.chainEnabled,
     },
     transport: {
-      bpm: compact.bpm ?? DEFAULT_BPM,
-      swing: compact.sw ?? DEFAULT_SWING,
+      bpm: data.bpm ?? INIT_DOCUMENT.transport.bpm,
+      swing: data.sw ?? INIT_DOCUMENT.transport.swing,
     },
-    sequencer: {
-      pattern,
-      variationCycle: (compact.vc ?? DEFAULT_VARIATION_CYCLE) as
-        | VariationCycle
-        | undefined,
-      chain,
-      chainEnabled,
-    },
-    masterChain: decodeMasterChain(compact.mc),
+    master: decodeMaster(
+      data.mc as CompactMaster | undefined,
+      INIT_DOCUMENT.master,
+    ),
   };
+
+  const result = presetDocumentSchema.safeParse(candidate);
+  if (!result.success) {
+    const issue = result.error.issues[0];
+    throw new CorruptFieldError(issue.path.join("."), issue.message);
+  }
+  return result.data;
 }
 
-export { encodeCompactPreset, decodeCompactPreset };
-export type { CompactPreset };
+export { COMPACT_CODEC_VERSION, decodeCompactDocument, encodeCompactDocument };
+export type { CompactChannel, CompactMaster, CompactPreset };
