+++ src/features/preset/lib/serialization/url-codec.test.ts
@@ -1,16 +1,17 @@
 /**
- * Share-URL codec: version dispatch, v1.5 fixture pinning, v2 round-trip
+ * Share-URL codec: latest-only version dispatch, canonical round-trip
  * tolerance, and typed decode failures.
  *
- * The v1.5 fixtures in __fixtures__/ are frozen payloads generated by the
- * pre-v2 codec (see __fixtures__/README.md); they prove that links shared
- * by v1.5 builds keep decoding through the retained v1.5 path.
+ * There is one share codec (compact.ts). A payload at the single current
+ * COMPACT_CODEC_VERSION decodes; every other `v` - older v1.5 links and
+ * versionless pre-#269 links - is refused with UnsupportedVersionError
+ * (#373). Saved `.dh` files still migrate old formats through the document
+ * ladder; that is exercised by the document tests, not here.
  */
 
 import { readFileSync } from "node:fs";
 import { describe, expect, it } from "vitest";
 
-import { init } from "@/core/dh";
 import {
   CorruptFieldError,
   decodePresetFileText,
@@ -21,18 +22,15 @@ import {
   type PresetDocument,
 } from "@/features/preset/document";
 import { buildDenseSharePreset } from "./__fixtures__/dense-preset";
-import { encodeCompactDocument, type CompactPresetV2 } from "./compact-v2";
+import {
+  COMPACT_CODEC_VERSION,
+  encodeCompactDocument,
+  type CompactPreset,
+} from "./compact";
 import { compress } from "./compress";
 import { shareableDocumentToUrl, urlToDocument } from "./index";
 
 // --- Helpers ----------------------------------------------------------------
-
-function shareFixture(name: string): string {
-  return readFileSync(
-    new URL(`./__fixtures__/${name}`, import.meta.url),
-    "utf-8",
-  ).trim();
-}
 
 function documentFixture(name: string): string {
   return readFileSync(
@@ -41,27 +39,25 @@ function documentFixture(name: string): string {
   );
 }
 
-/** Everything but the decode-minted timestamps. */
-function stripTimestamps(document: PresetDocument) {
-  const { createdAt: _c, updatedAt: _u, ...meta } = document.meta;
-  return { ...document, meta };
-}
-
 /**
- * Half the 0.1 knob display step: the codec's quantization contract
- * (see the precision table in compact-v2.ts).
+ * Assert two documents survive a share round trip within the codec's declared
+ * CANONICAL tolerance.
+ *
+ * The codec is deliberately lossy to save URL bytes: each canonical numeric
+ * field is quantized to a fixed number of decimals (compact.ts PRECISION), so
+ * a round trip perturbs a field by at most one grid step of 10^-p canonical
+ * units in that field's own unit (seconds, dB, Hz, semitones, 0..1 fraction).
+ * We compare each field against that canonical epsilon directly - there is no
+ * knob-space conversion anywhere in this assertion. Integer-carried fields
+ * (compRatio) and untouched fields (bpm, kit id, name, pattern structure,
+ * chain) are lossless and compared for exact equality; velocities share the
+ * pattern codec's 0-100 quantization (a 0.005 fraction epsilon).
  */
-/**
- * Assert two documents survive a v2 round trip within the codec's declared
- * precision: pattern, chain, kit, and identity data exact (velocities within
- * the shared 0-100 quantization), and every canonical numeric field within
- * one decimal grid step of compact-v2.ts PRECISION.
- */
-function expectKnobEquivalent(
+function expectCanonicalRoundTrip(
   actual: PresetDocument,
   expected: PresetDocument,
 ): void {
-  // Exact fields first.
+  // Lossless fields: exact equality.
   expect(actual.kit.id).toBe(expected.kit.id);
   expect(actual.meta.id).toBe(expected.meta.id);
   expect(actual.meta.name).toBe(expected.meta.name);
@@ -90,9 +86,8 @@ function expectKnobEquivalent(
     expected.pattern.variationMetadata,
   );
 
-  // Domain-unit comparison: the v2 codec quantizes each field to a fixed
-  // decimal precision (compact-v2.ts PRECISION), so a round trip perturbs it
-  // by at most half a grid step. Compare canonical values within one step.
+  // Lossy canonical fields: within one quantization grid step (10^-p) of the
+  // field's own canonical unit. `decimals` is the PRECISION for that field.
   const closeTo = (a: number, b: number, decimals: number, label: string) =>
     expect(Math.abs(a - b), label).toBeLessThanOrEqual(10 ** -decimals);
 
@@ -102,7 +97,7 @@ function expectKnobEquivalent(
       channel.decaySeconds,
       expectedChannel.decaySeconds,
       7,
-      `channel ${i} decay`,
+      `channel ${i} decay (s)`,
     );
     expect(channel.filter.side, `channel ${i} filter side`).toBe(
       expectedChannel.filter.side,
@@ -111,7 +106,7 @@ function expectKnobEquivalent(
       channel.filter.cutoffHz,
       expectedChannel.filter.cutoffHz,
       3,
-      `channel ${i} cutoff`,
+      `channel ${i} cutoff (Hz)`,
     );
     if (expectedChannel.volumeDb === null) {
       expect(channel.volumeDb, `channel ${i} volume`).toBeNull();
@@ -120,7 +115,7 @@ function expectKnobEquivalent(
         channel.volumeDb!,
         expectedChannel.volumeDb,
         2,
-        `channel ${i} volume`,
+        `channel ${i} volume (dB)`,
       );
     }
     closeTo(channel.pan, expectedChannel.pan, 4, `channel ${i} pan`);
@@ -128,7 +123,7 @@ function expectKnobEquivalent(
       channel.tuneSemitones,
       expectedChannel.tuneSemitones,
       3,
-      `channel ${i} tune`,
+      `channel ${i} tune (semitones)`,
     );
     expect(channel.solo, `channel ${i} solo`).toBe(expectedChannel.solo);
     expect(channel.mute, `channel ${i} mute`).toBe(expectedChannel.mute);
@@ -137,30 +132,40 @@ function expectKnobEquivalent(
   const am = actual.master;
   const em = expected.master;
   expect(am.filter.side, "master filter side").toBe(em.filter.side);
-  closeTo(am.filter.cutoffHz, em.filter.cutoffHz, 3, "master cutoff");
+  closeTo(am.filter.cutoffHz, em.filter.cutoffHz, 3, "master cutoff (Hz)");
   closeTo(am.saturation, em.saturation, 4, "master saturation");
   closeTo(am.phaser, em.phaser, 4, "master phaser");
   closeTo(am.reverb, em.reverb, 4, "master reverb");
-  closeTo(am.compThresholdDb, em.compThresholdDb, 2, "master compThreshold");
-  closeTo(am.compAttackSeconds, em.compAttackSeconds, 8, "master compAttack");
+  closeTo(
+    am.compThresholdDb,
+    em.compThresholdDb,
+    2,
+    "master compThreshold (dB)",
+  );
+  closeTo(
+    am.compAttackSeconds,
+    em.compAttackSeconds,
+    8,
+    "master compAttack (s)",
+  );
   closeTo(am.compMix, em.compMix, 4, "master compMix");
   if (em.masterVolumeDb === null) {
     expect(am.masterVolumeDb, "master volume").toBeNull();
   } else {
-    closeTo(am.masterVolumeDb!, em.masterVolumeDb, 2, "master volume");
+    closeTo(am.masterVolumeDb!, em.masterVolumeDb, 2, "master volume (dB)");
   }
 
   closeTo(actual.transport.swing, expected.transport.swing, 4, "swing");
 }
 
-/** A schema-valid dense synthetic document exercising v2-only spellings. */
+/** A schema-valid dense synthetic document exercising every codec spelling. */
 function buildDenseSyntheticDocument(): PresetDocument {
   const document = migrateV1ToDocument(buildDenseSharePreset());
   return {
     ...document,
     channels: document.channels.map((channel, i) => ({
       ...channel,
-      // Off-grid domain values that must round-trip within knob tolerance.
+      // Off-grid canonical values that must round-trip within tolerance.
       decaySeconds: 0.005 + i * 0.61234567891,
       filter: {
         side: i % 2 === 0 ? ("lowpass" as const) : ("highpass" as const),
@@ -202,53 +207,13 @@ function payloadOf(value: unknown): string {
   return compress(JSON.stringify(value));
 }
 
-function validV2Payload(): CompactPresetV2 {
+function validPayload(): CompactPreset {
   return encodeCompactDocument(buildDenseSyntheticDocument());
 }
 
-// --- v1.5 fixture pinning -----------------------------------------------------
+// --- Round-trips --------------------------------------------------------------
 
-describe("v1.5 share fixtures keep decoding", () => {
-  it("decodes the frozen init link to the migrated init document", () => {
-    const document = urlToDocument(shareFixture("share-v1_5-init.txt"));
-
-    expect(document.kit.id).toBe("kit-0");
-    expect(document.meta.name).toBe("init");
-    expect(document.transport.bpm).toBe(100);
-    expect(document.transport.swing).toBe(0);
-
-    const expected = init();
-    expect(stripTimestamps(document)).toEqual(stripTimestamps(expected));
-  });
-
-  it("decodes the frozen dense link to the pinned dense document", () => {
-    const document = urlToDocument(shareFixture("share-v1_5-dense.txt"));
-
-    // Spot pins independent of the migration machinery.
-    expect(document.kit.id).toBe("kit-3");
-    expect(document.meta.id).toBe("6f9b1c22-0d4e-4b9a-9c33-7d1a2e5b8f40");
-    expect(document.meta.name).toBe("Dense Fixture");
-    expect(document.transport.bpm).toBe(137);
-    // v1.5 swing knob 64 through the v1.5 interpretation: 0.64 * 0.375.
-    expect(document.transport.swing).toBeCloseTo(0.24, 10);
-    expect(document.channels[0].volumeDb).toBeNull();
-    expect(document.master.masterVolumeDb).toBeNull();
-    expect(document.master.compRatio).toBe(6);
-    expect(document.playback.chainEnabled).toBe(true);
-    expect(document.playback.chain.steps).toEqual([
-      { variation: 0, repeats: 2 },
-      { variation: 1, repeats: 1 },
-      { variation: 3, repeats: 3 },
-    ]);
-
-    const expected = migrateV1ToDocument(buildDenseSharePreset());
-    expect(stripTimestamps(document)).toEqual(stripTimestamps(expected));
-  });
-});
-
-// --- v2 round-trips -----------------------------------------------------------
-
-describe("v2 codec round-trips within knob resolution", () => {
+describe("share codec round-trips within its canonical quantization", () => {
   const eraFixtures = [
     "v1-current.json",
     "v1-legacy-cycle.json",
@@ -261,31 +226,24 @@ describe("v2 codec round-trips within knob resolution", () => {
     it(`round-trips the migrated document of ${fixture}`, () => {
       const original = decodePresetFileText(documentFixture(fixture));
       const decoded = urlToDocument(shareableDocumentToUrl(original));
-      expectKnobEquivalent(decoded, original);
+      expectCanonicalRoundTrip(decoded, original);
     });
   }
-
-  it("round-trips the migrated v1.5 share fixtures", () => {
-    for (const fixture of ["share-v1_5-init.txt", "share-v1_5-dense.txt"]) {
-      const original = urlToDocument(shareFixture(fixture));
-      const decoded = urlToDocument(shareableDocumentToUrl(original));
-      expectKnobEquivalent(decoded, original);
-    }
-  });
 
   it("round-trips a dense synthetic document (null volumes, max chain)", () => {
     const original = buildDenseSyntheticDocument();
     const decoded = urlToDocument(shareableDocumentToUrl(original));
 
-    expectKnobEquivalent(decoded, original);
+    expectCanonicalRoundTrip(decoded, original);
     expect(decoded.channels[0].volumeDb).toBeNull();
     expect(decoded.master.masterVolumeDb).toBeNull();
     expect(decoded.playback.chain.steps).toHaveLength(8);
   });
 
-  it("encodes the stable kit id, not a positional code", () => {
-    const compact = validV2Payload();
-    expect(compact.v).toBe(2);
+  it("encodes the stable kit id and the current codec version", () => {
+    const compact = validPayload();
+    expect(compact.v).toBe(3);
+    expect(compact.v).toBe(COMPACT_CODEC_VERSION);
     expect(compact.k).toBe("kit-3");
   });
 });
@@ -307,9 +265,35 @@ describe("urlToDocument typed failures", () => {
     );
   });
 
+  it("refuses an old v1.5 knob-space link with UnsupportedVersionError", () => {
+    // Post-#269 v1.5 links are no longer read (#373): the share codec is
+    // latest-only and carries no legacy decoder.
+    const v15 = {
+      v: 1.5,
+      id: "old",
+      k: "0",
+      n: "v1.5 Link",
+      ip: [{}, {}, {}, {}, {}, {}, {}, {}],
+      pt: [],
+    };
+    expect(() => urlToDocument(payloadOf(v15))).toThrow(
+      UnsupportedVersionError,
+    );
+  });
+
+  it("refuses the superseded, ambiguous v: 2 payload with UnsupportedVersionError", () => {
+    // v: 2 labeled two incompatible filter shapes across the unreleased #357
+    // epic (#368); the single codec refuses it uniformly rather than decoding.
+    expect(() => urlToDocument(payloadOf({ ...validPayload(), v: 2 }))).toThrow(
+      UnsupportedVersionError,
+    );
+  });
+
   it("refuses unknown future versions with UnsupportedVersionError", () => {
     expect(() =>
-      urlToDocument(payloadOf({ ...validV2Payload(), v: 3 })),
+      urlToDocument(
+        payloadOf({ ...validPayload(), v: COMPACT_CODEC_VERSION + 1 }),
+      ),
     ).toThrow(UnsupportedVersionError);
   });
 
@@ -320,7 +304,7 @@ describe("urlToDocument typed failures", () => {
   });
 
   it("rejects a short ip array with CorruptFieldError at ip", () => {
-    const payload = { ...validV2Payload(), ip: [{}, {}] };
+    const payload = { ...validPayload(), ip: [{}, {}] };
     expect(() => urlToDocument(payloadOf(payload))).toThrow(CorruptFieldError);
     try {
       urlToDocument(payloadOf(payload));
@@ -331,8 +315,8 @@ describe("urlToDocument typed failures", () => {
 
   it("rejects a short pt array with CorruptFieldError at pt", () => {
     const payload = {
-      ...validV2Payload(),
-      pt: validV2Payload().pt.slice(0, 3),
+      ...validPayload(),
+      pt: validPayload().pt.slice(0, 3),
     };
     try {
       urlToDocument(payloadOf(payload));
@@ -344,7 +328,7 @@ describe("urlToDocument typed failures", () => {
   });
 
   it("rejects a corrupt trigger bitmap with a step-level path", () => {
-    const payload = validV2Payload();
+    const payload = validPayload();
     payload.pt[2].a.t = "zzzz";
     try {
       urlToDocument(payloadOf(payload));
@@ -356,7 +340,7 @@ describe("urlToDocument typed failures", () => {
   });
 
   it("rejects an out-of-range velocity step index", () => {
-    const payload = validV2Payload();
+    const payload = validPayload();
     payload.pt[0].a.v = { "-1": 50 };
     try {
       urlToDocument(payloadOf(payload));
@@ -368,7 +352,7 @@ describe("urlToDocument typed failures", () => {
   });
 
   it("rejects a malformed chain string with CorruptFieldError at ch", () => {
-    const payload = { ...validV2Payload(), ch: "A2Q9" };
+    const payload = { ...validPayload(), ch: "A2Q9" };
     try {
       urlToDocument(payloadOf(payload));
       expect.unreachable("should have thrown");
@@ -379,7 +363,7 @@ describe("urlToDocument typed failures", () => {
   });
 
   it("rejects an unknown kit id with UnknownKitError", () => {
-    const payload = { ...validV2Payload(), k: "kit-99" };
+    const payload = { ...validPayload(), k: "kit-99" };
     try {
       urlToDocument(payloadOf(payload));
       expect.unreachable("should have thrown");
@@ -390,7 +374,7 @@ describe("urlToDocument typed failures", () => {
   });
 
   it("rejects out-of-range domain values via the schema with a document path", () => {
-    const payload = validV2Payload();
+    const payload = validPayload();
     payload.ip[0] = { ...payload.ip[0], v: 99 }; // volumeDb max is 4
     try {
       urlToDocument(payloadOf(payload));
@@ -399,23 +383,5 @@ describe("urlToDocument typed failures", () => {
       expect(error).toBeInstanceOf(CorruptFieldError);
       expect((error as CorruptFieldError).path).toBe("channels.0.volumeDb");
     }
-  });
-});
-
-// --- Size sanity ----------------------------------------------------------------
-
-describe("encoded size report", () => {
-  it("logs v1.5 vs v2 payload lengths (informational, no assertion)", () => {
-    const initV15 = shareFixture("share-v1_5-init.txt");
-    const denseV15 = shareFixture("share-v1_5-dense.txt");
-    const initV2 = shareableDocumentToUrl(init());
-    const denseV2 = shareableDocumentToUrl(
-      migrateV1ToDocument(buildDenseSharePreset()),
-    );
-
-    console.log(
-      `share payload lengths (chars): init v1.5=${initV15.length} v2=${initV2.length}; ` +
-        `dense v1.5=${denseV15.length} v2=${denseV2.length}`,
-    );
   });
 });
